requirements.txt
numpy==2.2.2
Pillow==11.1.0
streamlit==1.41.1
streamlit_drawable_canvas==0.9.3
tensorflow==2.17.0
